requirements.txt
streamlit
snowflake-connector-python
python-dotenv
pandas
plotly
streamlit-option-menu
requests